feat(sdk): create sdk package
- Move function from @exposition/core to @exposition/sdk
- Update build configuration to include the sdk package
- Add alias for the sdk package
- Add sdk package to the publish list
- Updated all other packages to import from the sdk package
- Updated package dependencies
- Add package to the documentation

closes #36
