+++ docs/src/integrations/vue-devtools.md
@@ -1,8 +1,1 @@
 <!-- @include: ../../../packages/integrations/vue-devtools/README.md -->
-
-:::warning ⚠️ AaaaAAaaaAAa - I'm not that quick
-`Promise<🙇‍♂️> => 'I will add a guide as soon as possible!'` <br>
-Meanwhile please checkout the [hacky playground example](https://github.com/h2xd/exposition/blob/main/playground/src/main.ts).
-:::
-
-<<< ../playground/src/main.ts#vue-devtools{5-13}
