+++ docs/src/packages/sdk.md
@@ -1,0 +1,1 @@
+<!-- @include: ../../../packages/sdk/README.md -->
