+++ packages/core/index.ts
@@ -1,3 +1,2 @@
-export * from './src/sdk'
 export * from './src/functions/Exposition'
 export * from './src/@types/Exposition.types'
