--- packages/core/src/sdk/index.ts
@@ -1,7 +1,0 @@
-export * from './@types/exposition'
-export * from './@types/scenario'
-export * from './createExpositionState'
-export * from './getExpositionValues'
-export * from './getInitialExpositionValues'
-export * from './resetExpositionValues'
-export * from './updateExpositionValues'
