+++ packages/sdk/index.ts
@@ -1,0 +1,8 @@
+export * from './src/@types/exposition'
+export * from './src/@types/scenario'
+
+export * from './src/functions/createExpositionState'
+export * from './src/functions/getExpositionValues'
+export * from './src/functions/getInitialExpositionValues'
+export * from './src/functions/resetExpositionValues'
+export * from './src/functions/updateExpositionValues'
