+++ packages/sdk/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@exposition/core",
+  "name": "@exposition/sdk",
   "version": "0.2.0",
-  "description": "Core functionality of the exposition library",
-  "homepage": "https://h2xd.github.io/exposition/packages/core.html",
+  "description": "Functions and other utils to work with the exposition library",
+  "homepage": "https://h2xd.github.io/exposition/packages/sdk.html",
   "keywords": [
     "exposition",
     "mock",
@@ -23,18 +23,5 @@
     "main": "./dist/index.cjs",
     "module": "./dist/index.mjs"
   },
-  "license": "MIT",
-  "devDependencies": {
-    "cross-env": "^7.0.3",
-    "esbuild": "^0.12.9",
-    "events": "^3.3.0",
-    "rollup": "^2.52.3",
-    "rollup-plugin-dts": "^4.2.2",
-    "rollup-plugin-esbuild": "^4.9.1",
-    "type-fest": "^2.17.0",
-    "typescript": "^4.7.3"
-  },
-  "dependencies": {
-    "@exposition/sdk": "workspace:*"
-  }
+  "license": "MIT"
 }
