+++ packages/sdk/rollup.config.js
@@ -1,0 +1,3 @@
+import { definePackageBuild } from '../../scripts/build'
+
+export default definePackageBuild()
