+++ packages/sdk/src/@types/scenario.ts
