+++ packages/sdk/src/functions/createExpositionState.spec.ts
