+++ packages/sdk/src/functions/resetExpositionValues.spec.ts
