+++ packages/sdk/src/functions/updateExpositionValues.spec.ts
