+++ packages/web/README.md
@@ -1,3 +1,3 @@
-# @exposition/web
+# 🌐 @exposition/web
 
 Add utils for handling web interactions, w/ `localStorage`, r/w from `URLComponents` etc.
